+++ Class/Library.java
@@ -1,54 +1,12 @@
 package Class;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import Class.Song;
+import Controller.Process;
 public class Library {
 
-    public List<Song> createlibraryOfSong() {
-        List<Song> songList = new ArrayList<Song>();
-        Song song1 = new Song("Sobrio","Maluma","Reggaeton","Sobrio.jpge","Sobrio.descripcion","03:55","08/07/2021");
-        Song song2 = new Song("De vuelta pa la vuelta","Daddy Yanke","Reggaeton","Devueltpalavuelta.jpge","Devueltpalavuelta.descripcion","03:51","10/12/2020");
-        Song song3 = new Song("Cada quien","Maluma","Reggaeton","Cadaquien.jpge","Cada.descripcion","03:39","07/12/2021");
-        Song song4 = new Song("Pepas","Farruko","Reggaeton","Pepas.jpge","Pepas.descripcion","04:54","06/08/2021");
-        Song song5 = new Song("Poblado","Natan","Reggaeton","Poblado.jpge","Poblado.descripcion","03:22","11/12/2020");
-        Song song6 = new Song("Ayer y hoy","Jean Carlos","Vallenato","AyerYHoy.jpge","AyerYHoy.descripcion","05:05","01/06/2020");
-        Song song7 = new Song("Que pena","Jean Carlos","Vallenato","QuePena.jpge","QuePena.descripcion","05:28","16/11/2021");
-        Song song8 = new Song("Suave brisa","Los Inquietos","Vallenato","SuaveBrisa.jpge","SuaveBrisa.descripcion","04:27","09/09/2018");
-        Song song9 = new Song("Recuerdame","Los Inquietos","Vallenato","Recuerdame.jpge","Recuerdame.descripcion","03:26","04/10/2021");
-        Song song10 = new Song("Despues de amar","Los inquietos","Vallenato","DespuesDeAmar.jpge","DespuesDeAmar.descripcion","05:22","10/10/2021");
-        Song song11 = new Song("Necesito","Eddie Santiago","Salsa","Necesito.jpge","Necesito.descripcion","04:49","25/11/2010");
-        Song song12 = new Song("Tu me haces falta","Eddie Santiago","Salsa","TuMeHacesFalta.jpge","TuMeHacesFalta.descripcion","04:51","28/08/2021");
-        Song song13 = new Song("Que precio tiene el cielo","Marc Anthony","Salsa","QuePrecioTieneElCielo.jpge","QuePrecioTieneElCielo.descripcion","05:06","03/01/2018");
-        Song song14 = new Song("No me acostumnbro","Rey Ruiz","Salsa","NoMeAcostumbro.jpge","NoMeAcostumbro.descripcion","05:14","24/11/2009");
-        Song song15 = new Song("Hasta el sol de hoy","Edgar Joel","Salsa","HastaElSolDeHoy.jpge","HastaElSolDeHoy.descripcion","04:44","21/12/2010");
-        Song song16 = new Song("Ven","Fonseca","Pop","Ven.jpge","Ven.descripcion","04:20","08/02/2019");
-        Song song17 = new Song("Para siempre","Kanny Garcia","Pop","ParaSiempre.jpge","ParaSiempre.descripcion","03:26","16/02/2018");
-        Song song18 = new Song("Un ratito","Andrés Cepeda","Pop","UnRatito.jpge","UnRatito.descripcion","03:50","16/10/2012");
-        Song song19 = new Song("Piel Canela","Andrés Cepeda","Pop","PielCanela.jpge","PielCanela.descripcion","02:48","19/09/2018");
-        Song song20 = new Song("Desde lejos","Santiago Cruz","Pop","DesdeLejos.jpge","DesdeLejos.descripcion","04:11","14/05/2012");
-        songList.add(song1);
-        songList.add(song2);
-        songList.add(song3);
-        songList.add(song4);
-        songList.add(song5);
-        songList.add(song6);
-        songList.add(song7);
-        songList.add(song8);
-        songList.add(song9);
-        songList.add(song10);
-        songList.add(song11);
-        songList.add(song12);
-        songList.add(song13);
-        songList.add(song14);
-        songList.add(song15);
-        songList.add(song16);
-        songList.add(song17);
-        songList.add(song18);
-        songList.add(song19);
-        songList.add(song20);
-        return songList;
+
+    public void CreateListSong(){
+        Process process= new Process();
+        process.createlibraryOfSong();
     }
